completed till chapter 8 concepts

+++ 01_Hello_World/db/db_user.py
@@ -3,6 +3,8 @@ from schemas import UserBase
 from db.models import DbUser
 from db.hash import Hash
 
+from fastapi import HTTPException, status
+
 def create_user(db: Session, request: UserBase):
   new_user = DbUser(
     username = request.username,
@@ -18,11 +20,18 @@ def get_all_users(db: Session):
   return db.query(DbUser).all()
 
 def get_user(db: Session, id: int):
-  return db.query(DbUser).filter(DbUser.id == id).first()
+  user = db.query(DbUser).filter(DbUser.id == id).first()
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail=f'User with id {id} not found.')
+  return user
 
 def update_user(db:Session, id: int, request: UserBase):
-  user = db.query(DbUser).filter(DbUser.id == id)
+  user = db.query(DbUser).filter(DbUser.id == id).first()
   # Handle any exceptions
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail=f'User with id {id} not found.')
   user.update({
     DbUser.username: request.username,
     DbUser.email: request.email,
@@ -34,6 +43,9 @@ def update_user(db:Session, id: int, request: UserBase):
 def delete_user(db: Session, id: int):
   user = db.query(DbUser).filter(DbUser.id == id).first()
   # Handle any exceptions
+  if not user:
+    raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail=f'User with id {id} not found.')
   db.delete(user)
   db.commit()
   return 'ok'
